another module has been added to bundle the completeitem functionaity
